requirements.txt
flask
python-dotenv
pymupdf
fasttext-wheel
huggingface_hub
transformers
torch
Pillow
requests
IndicTransToolkit
flask-cors
google-generativeai
